docs: Combine game crash without mods docs
The libtbb crash can happen even before the match starts like the more than 8 cores crash, or it can happen 20 turns in. So it makes sense to put them together. Also we'll probably have to add the Mesa 24 Iris crash.

Also the ldd output for libtbb seems to be unreliable, so use md5sum instead.